third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/inline-script-blocked-expected.txt

CONSOLE ERROR: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*". Either the 'unsafe-inline' keyword, a hash ('sha256-tfI8qh894SA5Ci6mTaIpM7Yc7ipVIuo6Z6xxEBkkQBA='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*". Either the 'unsafe-inline' keyword, a hash ('sha256-Oyx3ttLvegYcsNPXU1IceH1NLku4Or68MYQJkVE4Igk='), or a nonce ('nonce-...') is required to enable inline execution.

This test passes if it doesn't alert fail.
